Add plant into dict

+++ src/back/Plants.Application/Requests/DictsRequest.cs
@@ -3,4 +3,4 @@
 namespace Plants.Application.Requests;
 
 public record DictsRequest() : IRequest<DictsResult>;
-public record DictsResult(Dictionary<int, string> Groups, Dictionary<int, string> Regions, Dictionary<int, string> Soils);
+public record DictsResult(Dictionary<long, string> Groups, Dictionary<long, string> Regions, Dictionary<long, string> Soils);
